requirements.txt
python-dotenv
faiss-cpu
langchain
langchain-community
langchain-google-genai
pypdf
docx2txt
